Move `MerchantContainer.tsx`, `MerchantDetailsForm.tsx` and `schema.ts` from `src/app/merchant-contract` to `src/app/merchant-contract/components`
